Refactor user and auth domains into model/router/service layers
User domain (#39):
- Create models/user.py (UserResponse, BalanceResponse)
- Create services/user.py (all balance CRUD, trade queries, portfolio logic)
- Rewrite routers/users.py as thin router (zero DB queries)
- Deprecate core/balance_utils.py (re-exports from services/user.py)

Auth domain (#40):
- Create models/auth.py (GoogleAuthRequest, AdminGoogleAuthRequest, EmailRegisterRequest, EmailLoginRequest)
- Create services/auth.py (Google OAuth, email auth, JWT token CRUD, admin auth)
- Rewrite routers/auth.py as thin router (zero DB queries, zero httpx calls)

Shared:
- Create models/common.py (APIResponse, PaginatedResponse)
- Create services/__init__.py
- Update models/responses.py and models/requests.py with re-exports for backward compatibility
- Update routers/admin.py to import APIResponse from models/common

All 45 endpoints verified working. No behavior change.

Closes #39
Closes #40
