+++ backend/services/__init__.py
@@ -1,0 +1,6 @@
+"""
+Service layer for VegaExchange.
+
+Each domain has its own service file containing all business logic and DB operations.
+Routers call services — services call DB and engines.
+"""
